+++ nutree/mermaid.py
@@ -47,8 +47,8 @@ def _node_to_mermaid_flowchart_iter(
     add_root: bool = True,
     unique_nodes: bool = True,
     headers: Iterable[str] | None = None,
-    node_mapper: MermaidNodeMapperCallbackType | None | str = None,
-    edge_mapper: MermaidEdgeMapperCallbackType | None | str = None,
+    node_mapper: MermaidNodeMapperCallbackType | str | None = None,
+    edge_mapper: MermaidEdgeMapperCallbackType | str | None = None,  # pyright: ignore[reportRedeclaration]
 ) -> Iterator[str]:
     """Generate Mermaid formatted output line-by-line.